Conflicts:
	src/main/webapp/fileUpload.jsp

+++ src/main/java/com/eagle/community/news/dao/CommunityNewsDao.java
@@ -4,6 +4,6 @@ import com.eagle.community.dao.BaseDao;
 import com.eagle.community.news.entity.CommunityNews;
 
 
-public interface CommunityNewsDao extends BaseDao<CommunityNews,Integer>{
+public interface CommunityNewsDao extends BaseDao<CommunityNews,Long>{
 	
 }
